kfputils: Move compute_component_id
